requirements2.txt
alembic==1.13.1
async-timeout==4.0.3
blinker==1.7.0
cachelib==0.12.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email-validator==1.0.5
Flask==2.3.3
Flask-Login==0.6.2
Flask-Migrate==4.0.7
Flask-Session==0.6.0
Flask-SQLAlchemy==3.0.5
Flask-Validator==1.4.1
greenlet==3.0.3
idna==3.6
isbnlib==3.9.10
iso3166==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
py-moneyed==0.8.0
python-dotenv==1.0.1
pytz==2019.3
redis==5.0.3
requests==2.31.0
schwifty==2020.2.2
SQLAlchemy==2.0.28
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==2.3.7
